more flexible gitstatusd lookup
If GITSTATUS_DAEMON is an absolute path, use that.
Else if GITSTATUS_DAEMON is a valid command name, use that.
Else if GITSTATUS_DAEMON is not empty, try $gitstatus_dir/{usrbin,bin}/$GITSTATUS_DAEMON.
Else try $gitstatus_dir/{usrbin,bin}/gitstatusd-$os-$arch{,-static}

When trying different binaries, stop on the first binary that can be can be started
by the program loader. Keep trying only if the file doesn't exist, not executable,
or not startable by the program loader.