expose HEAD's commit message
Add two new new message fields to the gitstatusd's reply and
expose them in bash and zsh bindings:

- VCS_STATUS_COMMIT_ENCODING:
  Encoding of the HEAD's commit message. Empty value means UTF-8.
- VCS_STATUS_COMMIT_SUMMARY:
  The first paragraph of the HEAD's commit message as one line.

Commit summary is truncated to 256 bytes. This can be overridden
with --max-commit-summary-length/-z.

Note that truncation may result in invalid UTF-8.

Messages in exotic may not work because of SafePrint() in
gitstatusd. This can be fixed.