don't force checking for dirty submodules
`opts.ignore_submodules = GIT_SUBMODULE_IGNORE_DIRTY` is equivalent to
setting the `git status` flag `--ignore-submodules=dirty`, which takes
precedence per-submodule `ignore` settings in .gitmodules. This means
that `gitstatus` returns a different result than a default `git status`
call with no flags if there are any dirty submodules that are marked
`ignore = all` in .gitsubmodules.

This is confusing because `git status` and `gitstatusd` will be out of
sync and there's no obvious reason why.

This should only result in a behavior change for  projects with
submodules that set the .gitmodules config value
`submodule.<name>.ignore` to something besides `none` or `dirty`, but in
all cases it will ensure `git status` and `gitstatusd` are in agreement.